requirements.txt
flask
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-google-genai
python-dotenv
pypdf
docx2txt
pandas
numpy
requests
